[#9] Rename fixtures to factories
