+++ test/components/result_list.js
@@ -3,7 +3,7 @@ import ResultList from '../../src/components/ResultList';
 import ResultRow from '../../src/components/ResultRow';
 import assert from 'power-assert';
 import { shallow } from 'enzyme';
-import createResult from '../fixtures/result';
+import createResult from '../factories/result';
 
 describe('ResultList', () => {
   let results;
