+++ test/reducers.js
@@ -1,7 +1,7 @@
 import * as reducers from '../src/reducers';
 import * as actions from '../src/actions';
 import assert from 'power-assert';
-import createResult from './fixtures/result';
+import createResult from './factories/result';
 
 describe('Reducers', () => {
   describe('query()', () => {
