+++ test/utils/bookmark_manager.js
@@ -3,7 +3,7 @@ import * as ItemTypes from '../../src/constants/item_types';
 import assert from 'power-assert';
 import sinon from 'sinon';
 import createMockChromeObject from '../helpers/chrome';
-import createBookmark from '../fixtures/bookmark';
+import createBookmark from '../factories/bookmark';
 
 describe('BookmarkManager', () => {
   let bookmarks;
